requirements.txt
asgiref==3.8.1
attrs==25.4.0
Django==4.2.21
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
gunicorn==20.1.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
whitenoise==6.6.0
importlib_metadata==8.7.1
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Markdown==3.9
pillow==11.3.0
PyJWT==2.10.1
PyYAML==6.0.3
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.2.0
zipp==3.23.0
